+++ tests/test_sparsecs.py
@@ -8,7 +8,7 @@ import torch
 
 from xformers.components.attention import maybe_sparsify
 from xformers.components.attention._sputnik_sparse import _dense_to_sparse
-from xformers.components.attention.core import SparseCS, _create_random_sparsity
+from xformers.components.attention.core import _create_random_sparsity, SparseCS
 
 B = 2
 M = 16  # not a nice round number, on purpose