+++ xformers/benchmarks/benchmark_sp24.py
@@ -8,10 +8,10 @@ from typing import Tuple
 
 import torch
 import torch.nn.functional as F
-from torch import nn
-from utils import DTYPE2STR, benchmark_main_helper2, product_dict
 
 import xformers.ops as xops
+from torch import nn
+from utils import benchmark_main_helper2, DTYPE2STR, product_dict
 
 min_run_time = 0.5
 device = torch.device("cuda")