+++ xformers/benchmarks/benchmark_tiled_matmul.py
@@ -9,7 +9,7 @@ import itertools
 import torch
 from torch.utils import benchmark
 
-from xformers.benchmarks.utils import DTYPE2STR, benchmark_main_helper
+from xformers.benchmarks.utils import benchmark_main_helper, DTYPE2STR
 from xformers.ops.tiled_matmul import tiled_matmul
 
 min_run_time = 5