+++ xformers/components/__init__.py
@@ -11,12 +11,14 @@ from xformers.utils import import_all_modules
 
 from .attention import Attention, build_attention  # noqa
 from .input_projection import InputProjection, InputProjectionConfig  # noqa
-from .residual import NormalizationType  # noqa
-from .residual import PostNorm  # noqa
-from .residual import PreNorm  # noqa
-from .residual import RequiresWrappedInputs  # noqa
-from .residual import Residual  # noqa
-from .residual import ResidualNormStyle  # noqa
+from .residual import (  # noqa
+    NormalizationType,
+    PostNorm,
+    PreNorm,
+    RequiresWrappedInputs,
+    Residual,
+    ResidualNormStyle,
+)
 
 warnings.warn(
     "xformers.components is deprecated and is not maintained anymore. "