+++ xformers/ops/__init__.py
@@ -11,15 +11,15 @@ from .fmha import (
     AttentionOp,
     AttentionOpBase,
     LowerTriangularMask,
+    memory_efficient_attention,
+    memory_efficient_attention_backward,
+    memory_efficient_attention_forward,
+    memory_efficient_attention_forward_requires_grad,
     MemoryEfficientAttentionCkOp,
     MemoryEfficientAttentionCutlassFwdFlashBwOp,
     MemoryEfficientAttentionCutlassOp,
     MemoryEfficientAttentionFlashAttentionOp,
     MemoryEfficientAttentionSplitKCkOp,
-    memory_efficient_attention,
-    memory_efficient_attention_backward,
-    memory_efficient_attention_forward,
-    memory_efficient_attention_forward_requires_grad,
 )
 from .indexing import index_select_cat, scaled_index_add
 from .modpar_layers import ColumnParallelLinear, RowParallelLinear
@@ -35,12 +35,12 @@ from .sequence_parallel_fused_ops import (
 from .sp24 import Sparse24Tensor, sparsify24, sparsify24_like
 from .swiglu_op import (
     SwiGLU,
+    swiglu,
     SwiGLUEagerOp,
     SwiGLUFusedOp,
     SwiGLUOp,
     SwiGLUOpDispatch,
     SwiGLUPackedFusedOp,
-    swiglu,
 )
 from .tiled_matmul import tiled_matmul
 from .unbind import get_stack_strides, stack_or_none, unbind